+++ packages/node-core/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Fixed
+- Fixed issue Buffer within sandbox could not concat Buffer and Uint8Array, also injected `atob`
+
 ### Added
 - Expose option to add workerData to workers (#2372)